add redesigned greasy tests
